Update patch set 6
Patch Set 6:

(1 comment)

docker needed rust compiler for python cryptography lib to build (but not to run)- it is currently installed & uninstalled in a single step. I added it not so nicely via piping downloaded setup script to shell (apk has a version thats too old). I'm not sure if it shouldn't be switched to sth with verification of downloaded contents? LMK

Patch-set: 6