Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Thank you for this cleanup.

While it should be benign and a no-op, making these changes IS a code change. You should add copyright lines for your company to each changed file.



There is a JIRA ticket I'd definitely like you to write that your code changes pointed out to me, and that's on the database execute calls. Please respond to the comment with the JIRA ticket number.



You might want to consider making the change to remove extraneous parentheses on if and while statements, such as changing

if (envPytest == 'test'):

to

if envPytest == 'test':

If you'd rather do that in a separate submission, I understand. If so, please add the JIRA ticket in the comments as well.

Patch-set: 1
Reviewer: Gerrit User 105 <105@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1